+++ source/sprites/AnimFlatten.cpp
@@ -17,7 +17,7 @@ void AnimFlatten::Clear()
 
 RenderReturn AnimFlatten::Draw(const RenderParams& rp, int frame) const
 {
-	if (frame >= 0 && frame < m_frames.size()) {
+	if (frame >= 0 && frame < static_cast<int>(m_frames.size())) {
 		return m_frames[frame].Draw(rp);
 	} else {
 		return RENDER_NO_DATA;
@@ -26,7 +26,7 @@ RenderReturn AnimFlatten::Draw(const RenderParams& rp, int frame) const
 
 bool AnimFlatten::Update(const UpdateParams& up, int frame)
 {
-	if (frame >= 0 && frame < m_frames.size()) {
+	if (frame >= 0 && frame < static_cast<int>(m_frames.size())) {
 		return m_frames[frame].Update(up, NULL);
 	} else {
 		return false;
@@ -35,7 +35,7 @@ bool AnimFlatten::Update(const UpdateParams& up, int frame)
 
 void AnimFlatten::SetFrame(const UpdateParams& up, int frame)
 {
-	if (frame >= 0 && frame < m_frames.size()) {
+	if (frame >= 0 && frame < static_cast<int>(m_frames.size())) {
 		m_frames[frame].SetFrame(up, frame);
 	}
 }
