+++ source/sprites/DrawDownsample.cpp
@@ -37,7 +37,8 @@ RenderReturn DrawDownsample::Draw(const Sprite* spr, const RenderParams& rp, flo
 	sl::ShaderMgr::Instance()->FlushShader();
 
 	RenderScissor::Instance()->Disable();
-	RenderCtxStack::Instance()->Push(RenderContext(RT->WIDTH, RT->HEIGHT, RT->WIDTH, RT->HEIGHT));
+	RenderCtxStack::Instance()->Push(RenderContext(
+		static_cast<float>(RT->WIDTH), static_cast<float>(RT->HEIGHT), RT->WIDTH, RT->HEIGHT));
 
 	rt->Bind();
 	DrawSpr2RT(spr, rp, downsample);
